INTERN-42 | auth and fix

+++ resources/views/task/show.blade.php
@@ -8,6 +8,12 @@
     <div>{{$task->description}}</div>
     <div>{{$task->status}}</div>
 
-    <a href="{{ route('tasks.destroy', $task->uuid) }}" data-confirm="Вы уверены?" data-method="delete" rel="nofollow">Удалить</a>
-    <a href="{{ route('tasks.edit', $task->uuid) }}">Обновить статус</a>
+    <?php
+    if ($task->creator_uuid === auth()->user()->uuid) {
+        ?>
+        <a href="{{ route('tasks.destroy', $task->uuid) }}" data-confirm="Вы уверены?" data-method="delete" rel="nofollow">Удалить</a>
+        <a href="{{ route('tasks.edit', $task->uuid) }}">Обновить статус</a>
+    <?php
+    }
+    ?>
 @endsection
